Handling Return Case
In ARM function computing, it is possible that a function may not have explicit return instruction signs and instead continue measuring size for the next function. To handle this scenario, we can detect when a new function's prologue is started after we have already started the prologue for the initial function. In this way, we can infer that the initial function has completed and another function has begun.